+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - "quit" and "exit" commands (useful, when used behind a socket)
 - *systemd* configuration files for sending the authentication requests/responses over a socket
 - Use HTTP/1.1 persistent connections for higher throughput
+- `-t generic` (equivalent to `-t prosody`) is now default
 ### Fixed
 ### Changed